requirements.txt
asgiref==3.4.1
dj-static==0.0.6
Django==3.2.7
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.2
static3==0.7.0
